Feature/vue

+++ Delivery/internal/repositories/models/Orders.go
@@ -1,9 +1,9 @@
 package models
 
 type Order struct {
-	Id        int     `json:"id"`
+	Id        int64   `json:"id"`
 	Price     float64 `json:"price"`
-	UserId    int     `json:"userId"`
+	UserId    int64   `json:"userId"`
 	Address   string  `json:"address"`
 	CreatedAt string  `json:"created_at"`
 	UpdatedAt string  `json:"updated_at"`
